rename causality_test files
