+++ giottotime/causality/__init__.py
@@ -3,8 +3,8 @@ The :mod:`giottotime.causality` module deals with the causality tests for time
 series data.
 """
 
-from .shifted_linear_coefficient import ShiftedLinearCoefficient
-from .shifted_pearson_correlation import ShiftedPearsonCorrelation
+from .linear_coefficient import ShiftedLinearCoefficient
+from .pearson_correlation import ShiftedPearsonCorrelation
 
 
 __all__ = [
